+++ R/print.bm_bitmap.R
@@ -1,23 +1,48 @@
 #' Print bitmap objects
 #'
-#' `print.bm_bitmap()` prints a representation of bitmap objects to the terminal
+#' `print.bm_bitmap()` prints a representation of bitmap objects to the terminal.
+#' It is a wrapper around `as.character.bm_bitmap()` which converts bitmap objects
+#' to a character vector.
 #'
+#' @param x A `bm_bitmap()` object
+#' @param ... Further arguments passed to or from other methods.
+#' @param labels Character vector of the labels to use for each integer value i.e.
+#'               The first character for integer `0L`,
+#'               the second character for integer `1L`, and so on.
 #' @examples
-#'   plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
-#'   plus_sign[5L, 3:7] <- 1L
-#'   plus_sign[3:7, 5L] <- 1L
-#'   plus_sign <- bm_bitmap(plus_sign)
-#'   print(plus_sign, labels = c(" ", "#"))
+#'   font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
+#'   font <- read_hex(font_file)
+#'   bm_R <- font[[str2ucp("R")]]
+#'   print(bm_R, labels = c(" ", "#"))
 #'
-#'   space_glyph <- bm_bitmap(matrix(0L, nrow = 8L, ncol = 8L))
-#'   print(space_glyph, labels = ".")
+#'   bm_8 <- font[[str2ucp("8")]]
+#'   bm_8_with_border <- bm_extend(bm_extend(bm_8, left = 1L),
+#'                                 padding = 1L, value = 2L)
+#'   print(bm_8_with_border, labels = c(".", "@", "X"))
+#'
 #' @seealso [bm_bitmap()]
-#' @inheritParams as.character.bm_bitmap
-#' @return A character vector of the string representation invisibly.
+#' @return A character vector of the string representation (`print.bm_bitmap()` does this invisibly).
+#'         As a side effect `print.bm_bitmap()` prints out the string representation to the terminal.
 #' @usage \method{print}{bm_bitmap}(x, ..., labels = c("\u2591", "\u2588", "\u2592"))
 #' @export
 print.bm_bitmap <- function(x, ..., labels = c("\u2591", "\u2588", "\u2592")) {
     x <- as_bm_bitmap(x)
     cat(as.character(x, ..., labels = labels), sep = "\n")
     invisible(x)
+}
+
+#' @usage \method{as.character}{bm_bitmap}(x, ..., labels = c("\u2591", "\u2588", "\u2592"))
+#' @rdname print.bm_bitmap
+#' @export
+as.character.bm_bitmap <- function(x, ..., labels = c("\u2591", "\u2588", "\u2592")) {
+    x <- as_bm_bitmap(x)
+    stopifnot(max(x) <= length(labels))
+    l <- sapply(seq_len(nrow(x)),
+                function(i) bm_bitmap_row_to_string(x[i, ], labels))
+    rev(l)
+}
+
+bm_bitmap_row_to_string <- function(row, labels) {
+    l <- lapply(row, function(i) labels[i + 1L])
+    do.call(paste0, l)
 }