+++ tests/testthat/test-bm_list.R
@@ -23,3 +23,15 @@ test_that("bm_list()", {
     expect_equal(glyph_list, bm_list(glyph_list))
     expect_equal(glyph_list, as_bm_list(glyph_list))
 })
+
+test_that("as_bm_list()", {
+    # Test 'as_bm_list.character()'
+    font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
+    font <- read_hex(font_file)
+    bml <- as_bm_list("RSTATS", font = font)
+    bml <- bm_extend(bml, padding = 1L, value = 0L)
+    bml <- bm_extend(bml, padding = c(2L, 1L), value = 2L)
+    bm <- do.call(cbind, bml)
+
+    verify_output("txt/RSTATS.txt", print(bm, labels = c(" ", "#", "X")))
+})